requirements.txt
Flask==1.0.2
gunicorn>=19.9.0
privex-loghelper>=1.0.4
privex-helpers>=1.1.0
python-dotenv>=0.10.1
redis>=3.2.0
pika==1.0.0
django-getenv>=1.3.2
dnspython>=1.16.0

# Required for 'peersapp' BGP route viewer
CouchDB>=1.2
protobuf>=3.9.0
grpcio>=1.22.0
grpcio-tools>=1.22.0
cloudant>=2.12.0